fix and clean up get config fns
