+++ src/constants.rs
@@ -514,7 +514,6 @@ pub const DEFAULT_BATTERY_LAYOUT: &str = r#"
 "#;
 
 // Config and flags
-pub const DEFAULT_CONFIG_FILE_PATH: &str = "bottom/bottom.toml";
 
 // TODO: Eventually deprecate this, or grab from a file.
 pub const CONFIG_TEXT: &str = r#"# This is a default config file for bottom.  All of the settings are commented
